Move reference parser methods to dedicated class
ReferenceArea doesn't seem like such a natural place for it, considering
all possible variants.
